Refactored by review
