+++ detached-examples/multiuser-scenario/index.js
@@ -1,26 +1,90 @@
-const express = require('express');
+const createTestCafe = require('testcafe');
 
-const app     = express();
-const port    = 3000;
+class User {
+    constructor (name, fileName, browser) {
+        this.name     = name;
+        this.fileName = fileName;
+        this.browser  = browser;
 
-app.get('/', (req, res) => {
-  res.sendFile(__dirname + '/index.html');
-});
+        this.promiseOfInit = new Promise(resolve => {
+            this.confirmCurrentStep = resolve;
+        });
 
-let color = void 0;
+        this.expectedStageName = void 0;
+        this.runExpectedStage  = null;
 
-app.get('/save', (req, res) => {
-    if (color)
-        res.sendFile(__dirname + '/not-ok.html');
-    else {
-        color = req.query.color;
-        res.sendFile(__dirname + '/ok.html');
+        this._runTest();
     }
-});
 
-module.exports = new Promise(resolve => {
-    app.listen(port, () => {
-        resolve();
-        console.log(`Example app listening at http://localhost:${port}`);
-    });
-});
+    async _runTest () {
+        const testcafe = await createTestCafe('localhost', 0, 0);
+        const runner   = testcafe.createRunner();
+
+        await runner
+            .src(this.fileName)
+            .browsers(this.browser)
+            .run();
+
+        testcafe.close();
+    }
+
+    async runStage (stageName) {
+        if (stageName !== this.expectedStageName) {
+            throw new Error(`Another stage was expected:
+                expected: ${this.expectedStageName}
+                trying run: ${stageName}`);
+        }
+
+        this.runExpectedStage();
+
+        return new Promise(resolve => {
+            this.confirmCurrentStep = resolve;
+        });
+    }
+}
+
+const scenarios = new Map();
+
+class Scenario {
+    constructor (description) {
+        scenarios.set(description, this);
+
+        this.users = new Map();
+    }
+
+    createUser (name, fileName, browser) {
+        const user = new User(name, fileName, browser);
+
+        this.users.set(name, user);
+
+        return user.promiseOfInit;
+    }
+
+    initUser (name) {
+        const user = this.users.get(name);
+    
+        if (!user)
+            throw new Error(`The user named '${name}' does not exist`);
+    
+        return function stage (stageName) {
+            user.confirmCurrentStep(user);
+    
+            user.expectedStageName = stageName;
+    
+            return new Promise(resolve => {
+                user.runExpectedStage = resolve;
+            });
+        };
+    }
+}
+
+function getScenario (description) {
+    const scenario = scenarios.get(description);
+    
+    if (scenario)
+        return scenario;
+
+    throw new Error(`The scenario '${description}' does not exist`);
+}
+
+module.exports = { Scenario, getScenario };
